+++ src/engine/EntityComponentSystem/Entity.ts
@@ -122,6 +122,15 @@ export class Entity<KnownComponents extends Component = never> extends Class imp
   }
 
   /**
+   * Adds a tag to an entity
+   * @param tag
+   * @returns Entity
+   */
+  public addTag(tag: string) {
+    return this.addComponent(new TagComponent(tag));
+  }
+
+  /**
    * The types of the components on the Entity
    */
   public get types(): string[] {
@@ -247,7 +256,7 @@ export class Entity<KnownComponents extends Component = never> extends Class imp
    * Removes all children from this entity
    */
   public removeAll(): Entity {
-    this.children.forEach(c => {
+    this.children.forEach((c) => {
       this.remove(c);
     });
     return this;
@@ -396,6 +405,8 @@ export class Entity<KnownComponents extends Component = never> extends Class imp
 
   /**
    * Get a component by type with typecheck
+   *
+   * (Does not work on tag components, use .hasTag("mytag") instead)
    * @param type
    */
   public get<T extends Component>(type: ComponentCtor<T>): T {